[Test] Fix Ming V1 phase 4 tests and prune duplicates
